More refactoring, more general
