+++ src/main/java/qupath/fx/controls/tour/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Create a tour of a JavaFX graphical user interface, highlighting elements and providing explanations.
+ * <p>
+ * This can be used as an introductory guide for new users.
+ */
+package qupath.fx.controls.tour;
